requirements.txt
flask>=3.0,<4.0
crewai>=0.108.0,<1.0
litellm>=1.30.0,<2.0
python-dotenv>=1.0,<2.0
markdown>=3.5,<4.0
